+++ CHANGELOG.md
@@ -10,6 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the current market price column to the historical market data table of the admin control
+- Introduced filters (`dataSource` and `symbol`) in the accounts endpoint
+
+### Changed
+
+- Switched to the accounts endpoint in the holding detail dialog
+
+## 2.107.1 - 2024-09-12
+
+### Fixed
+
+- Fixed an issue in the activities filters that occurred during destructuring
 
 ## 2.107.0 - 2024-09-10